refactor: Move utils to internal/
