+++ internal/utils/string.go
