+++ internal/utils/string_test.go
@@ -3,7 +3,7 @@ package utils_test
 import (
 	"testing"
 
-	"github.com/caffeine-addictt/waku/cmd/utils"
+	"github.com/caffeine-addictt/waku/internal/utils"
 	"github.com/stretchr/testify/assert"
 )
