+++ pkg/config/styles.go
@@ -5,7 +5,7 @@ import (
 	"path"
 	"path/filepath"
 
-	"github.com/caffeine-addictt/waku/cmd/utils"
+	"github.com/caffeine-addictt/waku/internal/utils"
 	"github.com/caffeine-addictt/waku/internal/types"
 )
